backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.0
pandas==2.2.2
pyarrow==16.1.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.0
torch==2.3.1
statsmodels==0.14.2
httpx==0.27.0
wandb==0.19.11
